Create font system and add Unica
